introduce join order enumerator
